+++ spec/registration_spec.rb
@@ -202,6 +202,9 @@ describe 'openstack-identity::registration' do
       context 'with sql catalog backend' do
         before do
           node.set['openstack']['identity']['catalog']['backend'] = 'sql'
+          node.set['openstack']['endpoints']['identity-internal']['host'] = '127.0.0.2'
+          node.set['openstack']['endpoints']['identity-internal']['port'] = '5001'
+          node.set['openstack']['endpoints']['identity-internal']['path'] = '/v2.0'
         end
 
         it 'registers identity endpoint' do
@@ -213,7 +216,7 @@ describe 'openstack-identity::registration' do
             service_type: 'identity',
             endpoint_region: 'RegionOne',
             endpoint_adminurl: 'http://127.0.0.1:35357/v2.0',
-            endpoint_internalurl: 'http://127.0.0.1:35357/v2.0',
+            endpoint_internalurl: 'http://127.0.0.2:5001/v2.0',
             endpoint_publicurl: 'http://127.0.0.1:5000/v2.0'
           )
         end