+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of cookbook-openstack-ide
 ## 10.0.0
 * Upgrading to Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5
+* Fix the internal endpoint URL by using the InternalURL variable rather than AdminURL
 
 ## 9.3.1
 * Add support for a templated keystone-paste.ini